encoder: add nanocbor_put_tstrn()
Often it is useful to write only parts of a string or a non-null
terminated string (e.g. obtained from `nanocbor_get_tstr()`.

`nanocbor_put_tstr()` does not work for this as expects a terminating
NULL byte.

Introduce a nanocbor_put_tstrn() function that copies a user defined
number of bytes.